Suggestion for novelty and reproduction criteria
The novelty and reproduction criteria were not really actionable for a reviewer. The paper should state if and how it is novel, or if it reproduces previous results (and any improvements). I believe this also avoids the pitfall of having reviewers rejecting papers because "it is not novel enough" (which is a general review antipattern detailed in the [empirical standards](https://github.com/acmsigsoft/EmpiricalStandards/blob/master/Empirical%20Standards.pdf) (page 5).